Remove concern and create common base class
